Major enhancements toward version 2.0.0
- Migrate to webpack
- Remove compiled files from repo (npm install compiles)
- Remove deprecated `apply`
- Use CommonJS includes
- Use `equals` Node module
- Use `Array.isArray`, even IE supports it now
- Remove all duplicated code
- Add `postinstall` command to be called by `npm install`
- Update README accordingly
- Update CONTRIBUTING accordingly
