+++ bower.json
@@ -3,7 +3,7 @@
   "version": "1.2.2",
   "homepage": "https://github.com/Starcounter-Jack/JSON-Patch",
   "description": "Fast implementation of JSON-Patch (RFC-6902) with duplex (observe changes) capabilities",
-  "main": "src/json-patch-duplex.js",
+  "main": "dist/fast-json-patch.js",
   "keywords": ["json", "patch", "http", "rest"],
   "license": "MIT",
   "ignore": [
