+++ package.json
@@ -22,33 +22,34 @@
     "url": "http://www.starcounter.com/"
   },
   "license": "MIT",
-  "main": "src/json-patch-duplex.js",
-  "typings": "src/json-patch-duplex.d.ts",
+  "main": "src/duplex.js",
+  "typings": "src/duplex.d.ts",
   "engines": {
     "node": ">= 0.4.0"
   },
   "devDependencies": {
+    "babili-webpack-plugin": "^0.1.1",
     "benchmark": "^2.1.2",
     "chalk": "^1.1.3",
-    "grunt": "^0.4.5",
-    "grunt-bump": "^0.3.0",
-    "grunt-contrib-uglify": "~0.5.0",
     "jasmine": "^2.5.1",
     "jsdom": "^9.5.0",
     "jsonfile": "^2.3.1",
     "typescript": "~2.0.0",
-    "underscore": "^1.8.3"
+    "underscore": "^1.8.3",
+    "webpack": "^2.6.1"
   },
   "scripts": {
     "tsc": "tsc",
+    "postinstall": "tsc",
     "tsc-watch": "tsc -w",
-    "uglify": "grunt uglify",
     "test": "npm run test-core && npm run test-duplex",
     "test-duplex": "jasmine DUPLEX=yes JASMINE_CONFIG_PATH=test/jasmine.json",
     "test-core": "jasmine  DUPLEX=no JASMINE_CONFIG_PATH=test/jasmine.json test/spec/jsonPatchTestsSpec.js test/spec/coreSpec.js test/spec/validateSpec.js",
     "bench": "npm run bench-core && npm run bench-duplex",
     "bench-core": "node test/spec/coreBenchmark.js",
-    "bench-duplex": "node test/spec/coreBenchmark.js DUPLEX=yes && node test/spec/duplexBenchmark.js",
-    "bench-duplex-only": "node test/spec/duplexBenchmark.js"
+    "bench-duplex": "node test/spec/coreBenchmark.js DUPLEX=yes && node test/spec/duplexBenchmark.js"
+  },
+  "dependencies": {
+    "deep-equal": "^1.0.1"
   }
 }
