--- test/SpecRunnerDuplex.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Jasmine Spec Runner v2.5.0</title>
-
-  <link rel="shortcut icon" type="image/png" href="lib/jasmine-2.5.0/jasmine_favicon.png">
-  <link rel="stylesheet" href="lib/jasmine-2.5.0/jasmine.css">
-
-  <script src="lib/jasmine-2.5.0/jasmine.js"></script>
-  <script src="lib/jasmine-2.5.0/jasmine-html.js"></script>
-  <!-- <script src="lib/jasmine-2.5.0/boot.js"></script> -->
-  <script src="lib/jasmine-custom-boot.js"></script>
-
-
-  <!-- include Benchmark.js -->
-  <script src="lib/benchmark.js/vendor/lodash/dist/lodash.compat.js"></script>
-  <script src="lib/benchmark.js/vendor/platform.js/platform.js"></script>
-  <script src="lib/benchmark.js/benchmark.js"></script>
-  <script src="lib/benchmark_reporter.js"></script>
-  <link rel="stylesheet" type="text/css" href="lib/benchmark_reporter.css">
-
-  <!-- include source files here... -->
-
-  <!-- include spec files here... -->
-  <script src="../src/json-patch-duplex.js"></script>
-  <!-- Dynamically load https://github.com/json-patch/json-patch-tests test suite -->
-  <script src="spec/jsonPatchTestsSpec.js"></script>
-  <script src="spec/coreSpec.js"></script>
-  <script src="spec/validateSpec.js"></script>
-  <script src="spec/duplexSpec.js"></script>
-  <script src="spec/duplexBenchmark.js"></script>
-  <script src="spec/coreBenchmark.js"></script>
-
-</head>
-
-<body>
-</body>
-</html>
