+++ tsconfig.json
@@ -2,7 +2,7 @@
     "compilerOptions": {
         "target": "es5",
         "module": "commonjs",
-        "noEmitHelpers": true,
+        "noEmitHelpers": false,
         "declaration": true,
         "removeComments": false,
         /* emitting `use strict` made PuppetJS break */
